Move notes into api
