+++ routes/api/notes/index.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import passport from 'passport';
 
-import models from '../../models';
+import models from '../../../models';
 
 const router = express.Router();
